unify error mapping
